Bit of refactoring, code movement.
