+++ README.md
@@ -61,5 +61,5 @@ pytest .
 ## Run integration tests (with Docker)
 
 ```
-python3 ./tests/reproduce_previous_builds.py
+python3 ./integration_tests/reproduce_previous_builds.py
 ```
