+++ integration_tests/previous_builds.py
@@ -1,16 +1,5 @@
-import json
-import os
-import shutil
-import subprocess
-import sys
-import urllib.request
-from pathlib import Path
-from typing import Dict, List, Optional, Tuple
 
-downloads_folder = Path("./testdata/downloads").resolve()
-extracted_folder = Path("./testdata/input/extracted").resolve()
-parent_output_folder = Path("./testdata/output").resolve()
-cargo_target_dir = Path("./testdata/output/cargo_target_dir").resolve()
+from typing import Dict, List, Optional
 
 
 class PreviousBuild:
@@ -30,7 +19,7 @@ class PreviousBuild:
         self.docker_image = docker_image
 
 
-builds: List[PreviousBuild] = [
+previous_builds: List[PreviousBuild] = [
     PreviousBuild(
         name="a.1",
         project_archive_url="https://github.com/multiversx/mx-reproducible-contract-build-example-sc/archive/refs/tags/v0.1.5.zip",
@@ -272,106 +261,3 @@ builds: List[PreviousBuild] = [
         docker_image="multiversx/sdk-rust-contract-builder:v3.1.0"
     )
 ]
-
-
-def main(cli_args: List[str]):
-    shutil.rmtree(downloads_folder, ignore_errors=True)
-    shutil.rmtree(extracted_folder, ignore_errors=True)
-    shutil.rmtree(parent_output_folder, ignore_errors=True)
-
-    downloads_folder.mkdir(parents=True, exist_ok=True)
-    extracted_folder.mkdir(parents=True, exist_ok=True)
-    cargo_target_dir.mkdir(parents=True, exist_ok=True)
-
-    selected_builds = ["a.1", "a.2", "a.3", "b.1", "b.2", "b.3", "c.1", "c.2", "c.3", "c.4", "c.5", "d.1", "e.1"]
-
-    for build in builds:
-        if not build.name in selected_builds:
-            continue
-
-        print("Reproducing build", build.name, "...")
-
-        project_path, packaged_src_path = fetch_source_code(build)
-        output_folder = parent_output_folder / build.name
-        output_folder.mkdir(parents=True, exist_ok=True)
-
-        if project_path and build.project_path_adjustment:
-            project_path = project_path / build.project_path_adjustment
-
-        run_docker(project_path, packaged_src_path, build.contract_name, build.docker_image, output_folder)
-
-        artifacts_path = output_folder / "artifacts.json"
-        artifacts_json = artifacts_path.read_text()
-        artifacts = json.loads(artifacts_json)
-
-        for contract_name, expected_code_hash in build.expected_code_hashs.items():
-            print(f"For contract {contract_name}, expecting code hash {expected_code_hash} ...")
-            codehash = artifacts[contract_name]["codehash"]
-            if len(codehash) != 64:
-                # It's an older image, "artifacts.json" contains a path towards the code hash, instead of the actual code hash
-                codehash = Path(output_folder / contract_name / codehash).read_text().strip()
-
-            if codehash != expected_code_hash:
-                raise Exception(f"{build.name}: codehash mismatch for contract {contract_name}! Expected {expected_code_hash}, got {codehash}")
-            print("OK, codehash matches!", codehash)
-
-
-def fetch_source_code(build: PreviousBuild) -> Tuple[Optional[Path], Optional[Path]]:
-    print("Fetching source code for", build.name, "...")
-
-    if build.project_zip_url:
-        downloaded_archive = downloads_folder / f"{build.name}.zip"
-        extracted_project = extracted_folder / build.name
-        urllib.request.urlretrieve(build.project_zip_url, downloaded_archive)
-        shutil.unpack_archive(downloaded_archive, extracted_project)
-        return extracted_project, None
-
-    if build.packaged_src_url:
-        downloaded_packaged_src = downloads_folder / f"{build.name}.json"
-        urllib.request.urlretrieve(build.packaged_src_url, downloaded_packaged_src)
-        return None, downloaded_packaged_src
-
-    raise Exception("No source code provided")
-
-
-def run_docker(
-        project_path: Optional[Path],
-        packaged_src_path: Optional[Path],
-        contract_name: Optional[str],
-        image: str,
-        output_folder: Path,
-):
-    docker_mount_args: List[str] = ["--volume", f"{output_folder}:/output"]
-
-    if project_path:
-        docker_mount_args.extend(["--volume", f"{project_path}:/project"])
-
-    if packaged_src_path:
-        docker_mount_args.extend(["--volume", f"{packaged_src_path}:/packaged-src.json"])
-
-    docker_mount_args += ["--volume", f"{cargo_target_dir}:/rust/cargo-target-dir"]
-
-    docker_args = ["docker", "run"]
-    docker_args += docker_mount_args
-    docker_args += ["--user", f"{str(os.getuid())}:{str(os.getgid())}"]
-    docker_args += ["--rm", image]
-
-    entrypoint_args: List[str] = []
-
-    if project_path:
-        entrypoint_args.extend(["--project", "project"])
-
-    if packaged_src_path:
-        entrypoint_args.extend(["--packaged-src", "packaged-src.json"])
-
-    if contract_name:
-        entrypoint_args.extend(["--contract", contract_name])
-
-    args = docker_args + entrypoint_args
-
-    result = subprocess.run(args)
-    return result.returncode
-
-
-if __name__ == "__main__":
-    main(sys.argv[1:])
