+++ integration_tests/project_folder_and_packaged_src_are_equivalent.py
